luxy_home_prints-main/requirements.txt
asgiref==3.5.0
dj-database-url==1.0.0
Django==4.0.4
django-admin-interface==0.19.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
